refactor(forms): extract HasMinMax/HasStep for date-time inputs
DateInput, TimeInput, and DateTimeInput copy-pasted identical min/max (and step)
string setters; move them into shared concerns.
