+++ src/Forms/Components/Concerns/HasStep.php
@@ -1,0 +1,16 @@
+<?php
+declare(strict_types=1);
+
+namespace Lattice\Lattice\Forms\Components\Concerns;
+
+trait HasStep
+{
+    public ?int $step = null;
+
+    public function step(int $step): static
+    {
+        $this->step = $step;
+
+        return $this;
+    }
+}
