+++ src/Forms/Components/DateInput.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 namespace Lattice\Lattice\Forms\Components;
 
 use Lattice\Lattice\Forms\Attributes\AsField;
+use Lattice\Lattice\Forms\Components\Concerns\HasMinMax;
 use Lattice\Lattice\Forms\Enums\FieldType;
 use Lattice\Lattice\Ui\Concerns\HasAutoFocus;
 use Lattice\Lattice\Ui\Concerns\HasTabIndex;
@@ -12,23 +13,6 @@ use Lattice\Lattice\Ui\Concerns\HasTabIndex;
 class DateInput extends Field
 {
     use HasAutoFocus;
+    use HasMinMax;
     use HasTabIndex;
-
-    public ?string $min = null;
-
-    public ?string $max = null;
-
-    public function min(string $min): static
-    {
-        $this->min = $min;
-
-        return $this;
-    }
-
-    public function max(string $max): static
-    {
-        $this->max = $max;
-
-        return $this;
-    }
 }
