+++ src/Forms/Components/TimeInput.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 namespace Lattice\Lattice\Forms\Components;
 
 use Lattice\Lattice\Forms\Attributes\AsField;
+use Lattice\Lattice\Forms\Components\Concerns\HasMinMax;
+use Lattice\Lattice\Forms\Components\Concerns\HasStep;
 use Lattice\Lattice\Forms\Enums\FieldType;
 use Lattice\Lattice\Forms\Rules\TimeString;
 use Lattice\Lattice\Ui\Concerns\HasAutoFocus;
@@ -13,34 +15,9 @@ use Lattice\Lattice\Ui\Concerns\HasTabIndex;
 class TimeInput extends Field
 {
     use HasAutoFocus;
+    use HasMinMax;
+    use HasStep;
     use HasTabIndex;
-
-    public ?string $min = null;
-
-    public ?string $max = null;
-
-    public ?int $step = null;
-
-    public function min(string $min): static
-    {
-        $this->min = $min;
-
-        return $this;
-    }
-
-    public function max(string $max): static
-    {
-        $this->max = $max;
-
-        return $this;
-    }
-
-    public function step(int $step): static
-    {
-        $this->step = $step;
-
-        return $this;
-    }
 
     /**
      * @return array<int, mixed>
